updated autodetection at open and enumerate

+++ client/xibridge_client.cpp
@@ -121,7 +121,7 @@ bool Xibridge_client::exec_enumerate(
         int num_devs = proto->get_result_error();
         if (num_devs != 0)
         {
-            MBuf res((strlen(_host) + sizeof(xibridge_device_t) * 2 + 16/* xi-net://<>/<>*/) * num_devs);
+            MBuf res(((int)(strlen(_host) + sizeof(xibridge_device_t)) * 2 + 16/* xi-net://<>/<>*/) * num_devs);
             xi_net_dev_uris(res, _host, data, num_devs);
             *result = (char *)malloc(res.realSize());
             memcpy(*result, (const uint8_t *)res, res.realSize());
